requirements.txt
fire==0.5.0
pandas==2.0.3
python_docx==1.1.0
scikit_learn==1.3.2
Pillow
PyMuPDF
PyPDF2
pdf2docx
docx2pdf
pytesseract
openai==0.28.1
icalendar
